test: move os testes de usuário para a pasta só de teste
